requirements.txt
uvicorn>=0.16.0,<0.18.0
fastapi>=0.64.0,<1.0.0
sqlalchemy>=1.4.37,<1.4.40
databases==0.6.1
pydantic[dotenv]<2.0.0,>=1.6.2
aiomysql>=0.0.21,<0.1.1
mysqlclient>=2.0.0,<2.1.1
PyMySQL>=0.93,<=1.0.2
asyncpg>=0.22.0,<0.26.0
psycopg2==2.9.1
passlib>=1.6.4,<=1.7.4
bcrypt>=3.1.5,<=3.2.2
python-multipart>=0.0.1,<0.0.5
PyJWT>=2.0.0,<=2.4.0
fastapi-mail>=1.0.6,<=1.1.4
aioredis==2.0.1
httpx==0.23.0
# starlette
# pytz
# pydantic[dotenv]
# asyncpg
# psycopg2
# passlib
# pyjwt
# python-jose[cryptography]
# bcrypt
# python-multipart
# fastapi-mail
# Pillow
